2 eme commit

+++ libft/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: alsomvil <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/06 04:53:53 by alsomvil          #+#    #+#             */
-/*   Updated: 2018/02/21 11:14:11 by alsomvil         ###   ########.fr       */
+/*   Updated: 2017/12/06 04:54:57 by alsomvil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(long n)
+void	ft_putnbr(int n)
 {
 	ft_putnbr_fd(n, 1);
 }
